# src/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  version: 1.0.0
  description: A simple Todo app API with bearer-token authentication.

servers:
  - url: https://api.example.com/v1

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - completed
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Buy milk"
        description:
          type: string
          example: "2% milk, one gallon"
        completed:
          type: boolean
          example: false
        dueDate:
          type: string
          format: date-time
          example: "2025-09-03T12:00:00Z"
    TodoArray:
      type: array
      items:
        $ref: '#/components/schemas/Todo'

    NewTodo:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date-time

    UpdateTodo:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        dueDate:
          type: string
          format: date-time

security:
  - bearerAuth: []

paths:
  /todos:
    get:
      summary: List all todos
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoArray'

    post:
      summary: Create a new todo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTodo'
      responses:
        "201":
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'

  /todos/{id}:
    get:
      summary: Get a todo by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        "404":
          description: Todo not found

    put:
      summary: Update a todo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodo'
      responses:
        "200":
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        "404":
          description: Todo not found

    delete:
      summary: Delete a todo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Todo deleted
        "404":
          description: Todo not found